Play Merlob vs. PG
